fix: patch 'use' command to enable user to switch between git profiles (#18)
* feat: add types + constants

* feat: add profile store + active store

* feat: store backup + file lock - safe file writes

* feat: add atomic write + directory setup

* feat: add profile name validations

* fix: add fsync to atomicWrite

* feat: add config store

* feat: add utils - output formatter + cli prompter

* feat: implement add command + unit tests

* feat: implement ls command + unit tests

* feat: implement remove command + unit tests

* feat: implement show command + unit tests

* feat: implement use command + unit tests - switch profiles

* feat: get current active profile

* feat: add all commands to cli.ts - yargs

* feat: add unit test cases for utils - output + profileNameValidation

* docs: update README.md

* refactor: remove package-lock.json

* refactor: update bun-lock + root README.md

* chore: add CI workflow + test coverage threshold

* fix: patch use command to enable profile switching

* refactor: remove types + interfaces to separate files

* refactor: remove redundant lines of code

---------
