+++ .husky/pre-commit
@@ -1,1 +1,3 @@
 bunx lint-staged
+bun run build
+bun run test:coverage
