+++ src/core/profileManagement/profiles.ts
@@ -1,4 +1,4 @@
-import type { Profile } from '../../lib/types';
+import type { Profile, ProfilesStore } from '../../lib/types/Profile.type';
 import { ExitCode } from '../../lib/constants';
 import { loadProfiles, saveProfiles } from './profileStore';
 import { ProfileError } from './errorClass';
@@ -7,8 +7,8 @@ import { validateProfileName } from './validateProfileName';
 
 // Profile CRUD
 const getProfile = (name: string): Profile | undefined => {
-  const store = loadProfiles();
-  return store.profiles.find(p => p.name === name);
+  const store: ProfilesStore = loadProfiles();
+  return store.profiles.find((p) => p.name === name);
 };
 
 const listProfiles = (): Profile[] => {
@@ -16,8 +16,8 @@ const listProfiles = (): Profile[] => {
 };
 
 const addProfile = async (profile: Profile): Promise<void> => {
-  const store = loadProfiles();
-  const exists = store.profiles.some(p => p.name === profile.name);
+  const store: ProfilesStore = loadProfiles();
+  const exists = store.profiles.some((p) => p.name === profile.name);
   if (exists) {
     throw new ProfileError('profile already exists', ExitCode.PROFILE_ALREADY_EXISTS);
   }
@@ -27,37 +27,37 @@ const addProfile = async (profile: Profile): Promise<void> => {
 };
 
 const updateProfile = async (name: string, updates: Partial<Profile>): Promise<void> => {
-  const store = loadProfiles();
+  const store: ProfilesStore = loadProfiles();
 
   if (updates.name) {
-    const exists = store.profiles.some(p => p.name === updates.name && p.name !== name);
+    const exists = store.profiles.some((p) => p.name === updates.name && p.name !== name);
     if (exists) {
       throw new ProfileError('profile already exists', ExitCode.PROFILE_ALREADY_EXISTS);
     }
   }
 
-  const index = store.profiles.findIndex(p => p.name === name);
+  const index = store.profiles.findIndex((p) => p.name === name);
   if (index === -1) {
     throw new ProfileError('profile not found', ExitCode.PROFILE_NOT_FOUND);
   }
 
   store.profiles[index] = {
     ...store.profiles[index],
-    ...updates
-  } as Profile
+    ...updates,
+  } as Profile;
 
   await saveProfiles(store);
 };
 
 const removeProfile = async (name: string, force?: boolean): Promise<void> => {
-  const store = loadProfiles();
+  const store: ProfilesStore = loadProfiles();
   const active = getActiveProfileName();
 
   if (active === name && !force) {
     throw new ProfileError('cannot remove active profile. Use --force', ExitCode.INVALID_INPUT);
   }
 
-  const updatedProfiles = store.profiles.filter(p => p.name !== name);
+  const updatedProfiles = store.profiles.filter((p) => p.name !== name);
 
   if (updatedProfiles.length === store.profiles.length) {
     throw new ProfileError('profile not found', ExitCode.PROFILE_NOT_FOUND);
@@ -72,11 +72,4 @@ const getActiveProfileName = (): string | null => {
   return active.global;
 };
 
-export {
-  validateProfileName,
-  getProfile,
-  listProfiles,
-  addProfile,
-  updateProfile,
-  removeProfile
-}
+export { validateProfileName, getProfile, listProfiles, addProfile, updateProfile, removeProfile };
