+++ src/lib/types/ActiveStore.type.ts
@@ -1,0 +1,4 @@
+export interface ActiveStore {
+  global: string | null;
+  switched_at: string | null;
+}
