+++ src/lib/types/AddCommand.type.ts
@@ -1,0 +1,10 @@
+export interface AddArgs {
+  name: string;
+  displayName?: string;
+  email?: string;
+  sshKey?: string;
+  gpgKey?: string;
+  signing?: boolean;
+  noInteractive?: boolean;
+  json?: boolean;
+}
