+++ src/lib/types/JsonOutput.type.ts
@@ -1,0 +1,5 @@
+export interface JsonOutput<T = unknown> {
+  success: boolean;
+  data?: T;
+  error?: { code: number; message: string };
+}
