+++ vitest.config.ts
@@ -8,7 +8,13 @@ export default defineConfig({
         clearMocks: true,
         coverage: {
             provider: 'v8',
-            reporter: ['lcov', 'html', 'text']
+            reporter: ['lcov', 'html', 'text'],
+            thresholds: {
+                statements: 80,
+                branches: 75,
+                functions: 75,
+                lines: 80
+            }
         },
         globals: true
     },
